+++ addOns/requester/src/main/java/org/zaproxy/zap/extension/requester/ManualHttpRequestEditorPanel.java
@@ -40,6 +40,7 @@ import org.apache.commons.httpclient.URIException;
 import org.apache.log4j.Logger;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.model.Model;
+import org.parosproxy.paros.model.OptionsParam;
 import org.parosproxy.paros.network.HttpHeader;
 import org.parosproxy.paros.network.HttpMalformedHeaderException;
 import org.parosproxy.paros.network.HttpMessage;
@@ -134,7 +135,7 @@ public class ManualHttpRequestEditorPanel extends ManualRequestEditorPanel {
     }
 
     @Override
-    protected MessageSender getMessageSender() {
+    protected HttpPanelSender getMessageSender() {
         return sender;
     }
 
@@ -300,7 +301,7 @@ public class ManualHttpRequestEditorPanel extends ManualRequestEditorPanel {
         try {
             URI uri = new URI("http://www.any_domain_name.org/path", true);
             msg.setRequestHeader(
-                    new HttpRequestHeader(HttpRequestHeader.GET, uri, HttpHeader.HTTP10));
+                    new HttpRequestHeader(HttpRequestHeader.GET, uri, HttpHeader.HTTP11));
             setMessage(msg);
         } catch (HttpMalformedHeaderException e) {
             logger.error(e.getMessage(), e);
@@ -636,16 +637,22 @@ public class ManualHttpRequestEditorPanel extends ManualRequestEditorPanel {
     }
 
     public void addPersistentConnectionListener(PersistentConnectionListener listener) {
-        ((HttpPanelSender) getMessageSender()).addPersistentConnectionListener(listener);
+        getMessageSender().addPersistentConnectionListener(listener);
     }
 
     public void removePersistentConnectionListener(PersistentConnectionListener listener) {
-        ((HttpPanelSender) getMessageSender()).removePersistentConnectionListener(listener);
+        getMessageSender().removePersistentConnectionListener(listener);
     }
 
     public void beforeClose() {
         HttpPanelManager panelManager = HttpPanelManager.getInstance();
         panelManager.removeRequestPanel(getRequestPanel());
         panelManager.removeResponsePanel(getResponsePanel());
+    }
+
+    void optionsChanged(OptionsParam optionsParam) {
+        getMessageSender()
+                .setButtonTrackingSessionStateEnabled(
+                        optionsParam.getConnectionParam().isHttpStateEnabled());
     }
 }