+++ addOns/requester/CHANGELOG.md
@@ -7,9 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Add help.
 - Add info and repo URLs.
+- Allow to disable cookies (Issue 4934).
 
 ### Changed
 - Update minimum ZAP version to 2.9.0.
+
+### Fixed
+- Add the requests to the Sites tree to be able to active scan them (Issue 5778).
+- Enforce the mode when sending the request and following redirections.
 
 ## 3 - 2018-10-15